creates `add_keys` mod and moves `scan_until()` function there
