+++ zebra-scan/src/service/scan_task/scan.rs
@@ -50,6 +50,8 @@ pub type State = Buffer<
     zebra_state::Request,
 >;
 
+mod add_keys;
+
 /// Wait a few seconds at startup for some blocks to get verified.
 ///
 /// But sometimes the state might be empty if the network is slow.
@@ -58,7 +60,7 @@ const INITIAL_WAIT: Duration = Duration::from_secs(15);
 /// The amount of time between checking for new blocks and starting new scans.
 ///
 /// This is just under half the target block interval.
-const CHECK_INTERVAL: Duration = Duration::from_secs(30);
+pub const CHECK_INTERVAL: Duration = Duration::from_secs(30);
 
 /// We log an info log with progress after this many blocks.
 const INFO_LOG_INTERVAL: u32 = 10_000;
@@ -122,7 +124,7 @@ pub async fn start(
                     Some((height, new_keys, state, storage)) = scan_until_task_receiver.recv() => {
                         // TODO: Add a long timeout?
                         let scan_until_task = tokio::spawn(
-                            scan_until(height, new_keys, state, storage).in_current_span(),
+                            add_keys::scan_until(height, new_keys, state, storage).in_current_span(),
                         );
 
                         scan_until_tasks.push(scan_until_task);
@@ -187,70 +189,6 @@ pub async fn start(
             .next()
             .expect("a valid blockchain never reaches the max height");
     }
-}
-
-/// Start a scan task that reads blocks from `state` within the provided height range,
-/// scans them with the configured keys in `storage`, and then writes the results to `storage`.
-pub async fn scan_until(
-    stop_before_height: Height,
-    keys: HashMap<SaplingScanningKey, (Vec<DiversifiableFullViewingKey>, Vec<SaplingIvk>, Height)>,
-    state: State,
-    storage: Storage,
-) -> Result<(), BoxError> {
-    let sapling_activation_height = storage.min_sapling_birthday_height();
-    // Do not scan and notify if we are below sapling activation height.
-    wait_for_height(
-        sapling_activation_height,
-        "Sapling activation",
-        state.clone(),
-    )
-    .await?;
-
-    let key_heights = keys
-        .iter()
-        .map(|(key, (_, _, height))| (key.clone(), *height))
-        .collect();
-    let key_heights = Arc::new(key_heights);
-
-    let mut height = get_min_height(&key_heights).unwrap_or(sapling_activation_height);
-
-    // Parse and convert keys once, then use them to scan all blocks.
-    let parsed_keys: HashMap<
-        SaplingScanningKey,
-        (Vec<DiversifiableFullViewingKey>, Vec<SaplingIvk>),
-    > = keys
-        .into_iter()
-        .map(|(key, (decoded_dfvks, decoded_ivks, _h))| (key, (decoded_dfvks, decoded_ivks)))
-        .collect();
-
-    let parsed_keys = Arc::new(parsed_keys);
-
-    // Give empty states time to verify some blocks before we start scanning.
-    tokio::time::sleep(INITIAL_WAIT).await;
-
-    while height < stop_before_height {
-        let scanned_height = scan_height_and_store_results(
-            height,
-            state.clone(),
-            None,
-            storage.clone(),
-            key_heights.clone(),
-            parsed_keys.clone(),
-        )
-        .await?;
-
-        // If we've reached the tip, sleep for a while then try and get the same block.
-        if scanned_height.is_none() {
-            tokio::time::sleep(CHECK_INTERVAL).await;
-            continue;
-        }
-
-        height = height
-            .next()
-            .expect("a valid blockchain never reaches the max height");
-    }
-
-    Ok(())
 }
 
 /// Polls state service for tip height every [`CHECK_INTERVAL`] until the tip reaches the provided `tip_height`
