+++ zebra-scan/src/service/scan_task/scan/add_keys.rs
@@ -1,0 +1,77 @@
+//! Functions for registering new keys in the scan task
+
+use std::{collections::HashMap, sync::Arc};
+
+use zcash_primitives::{sapling::SaplingIvk, zip32::DiversifiableFullViewingKey};
+use zebra_chain::{block::Height, BoxError};
+use zebra_state::SaplingScanningKey;
+
+use crate::{
+    scan::{scan_height_and_store_results, wait_for_height, State, CHECK_INTERVAL},
+    storage::Storage,
+};
+
+/// Start a scan task that reads blocks from `state` within the provided height range,
+/// scans them with the configured keys in `storage`, and then writes the results to `storage`.
+pub async fn scan_until(
+    stop_before_height: Height,
+    keys: HashMap<SaplingScanningKey, (Vec<DiversifiableFullViewingKey>, Vec<SaplingIvk>, Height)>,
+    state: State,
+    storage: Storage,
+) -> Result<(), BoxError> {
+    let sapling_activation_height = storage.min_sapling_birthday_height();
+    // Do not scan and notify if we are below sapling activation height.
+    wait_for_height(
+        sapling_activation_height,
+        "Sapling activation",
+        state.clone(),
+    )
+    .await?;
+
+    let key_heights: HashMap<String, Height> = keys
+        .iter()
+        .map(|(key, (_, _, height))| (key.clone(), *height))
+        .collect();
+    let key_heights = Arc::new(key_heights);
+
+    let mut height = key_heights
+        .values()
+        .cloned()
+        .min()
+        .unwrap_or(sapling_activation_height);
+
+    // Parse and convert keys once, then use them to scan all blocks.
+    let parsed_keys: HashMap<
+        SaplingScanningKey,
+        (Vec<DiversifiableFullViewingKey>, Vec<SaplingIvk>),
+    > = keys
+        .into_iter()
+        .map(|(key, (decoded_dfvks, decoded_ivks, _h))| (key, (decoded_dfvks, decoded_ivks)))
+        .collect();
+
+    let parsed_keys = Arc::new(parsed_keys);
+
+    while height < stop_before_height {
+        let scanned_height = scan_height_and_store_results(
+            height,
+            state.clone(),
+            None,
+            storage.clone(),
+            key_heights.clone(),
+            parsed_keys.clone(),
+        )
+        .await?;
+
+        // If we've reached the tip, sleep for a while then try and get the same block.
+        if scanned_height.is_none() {
+            tokio::time::sleep(CHECK_INTERVAL).await;
+            continue;
+        }
+
+        height = height
+            .next()
+            .expect("a valid blockchain never reaches the max height");
+    }
+
+    Ok(())
+}
